fix : 캘린더 특정 기기에서 잘리는 오류 해결
