--- KeKi/Scenes/Store/StoreImageCell.xib
@@ -1,39 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<document type="com.apple.InterfaceBuilder3.CocoaTouch.XIB" version="3.0" toolsVersion="21507" targetRuntime="iOS.CocoaTouch" propertyAccessControl="none" useAutolayout="YES" useTraitCollections="YES" useSafeAreas="YES" colorMatched="YES">
-    <device id="retina6_12" orientation="portrait" appearance="light"/>
-    <dependencies>
-        <deployment identifier="iOS"/>
-        <plugIn identifier="com.apple.InterfaceBuilder.IBCocoaTouchPlugin" version="21505"/>
-        <capability name="Safe area layout guides" minToolsVersion="9.0"/>
-        <capability name="documents saved in the Xcode 8 format" minToolsVersion="8.0"/>
-    </dependencies>
-    <objects>
-        <placeholder placeholderIdentifier="IBFilesOwner" id="-1" userLabel="File's Owner"/>
-        <placeholder placeholderIdentifier="IBFirstResponder" id="-2" customClass="UIResponder"/>
-        <collectionViewCell opaque="NO" clipsSubviews="YES" multipleTouchEnabled="YES" contentMode="center" id="gTV-IL-0wX" customClass="StoreImageCell" customModule="KeKi" customModuleProvider="target">
-            <rect key="frame" x="0.0" y="0.0" width="105" height="105"/>
-            <autoresizingMask key="autoresizingMask"/>
-            <view key="contentView" opaque="NO" clipsSubviews="YES" multipleTouchEnabled="YES" contentMode="center">
-                <rect key="frame" x="0.0" y="0.0" width="105" height="105"/>
-                <autoresizingMask key="autoresizingMask" flexibleMaxX="YES" flexibleMaxY="YES"/>
-                <subviews>
-                    <imageView clipsSubviews="YES" userInteractionEnabled="NO" contentMode="scaleAspectFit" horizontalHuggingPriority="251" verticalHuggingPriority="251" translatesAutoresizingMaskIntoConstraints="NO" id="TSG-Uf-nBd">
-                        <rect key="frame" x="0.0" y="0.0" width="105" height="105"/>
-                        <color key="backgroundColor" red="0.9882352941176471" green="0.95686274509803915" blue="0.87450980392156863" alpha="0.84705882352941175" colorSpace="calibratedRGB"/>
-                    </imageView>
-                </subviews>
-            </view>
-            <viewLayoutGuide key="safeArea" id="ZTg-uK-7eu"/>
-            <constraints>
-                <constraint firstAttribute="trailing" secondItem="TSG-Uf-nBd" secondAttribute="trailing" id="0hf-ql-R1d"/>
-                <constraint firstItem="TSG-Uf-nBd" firstAttribute="leading" secondItem="gTV-IL-0wX" secondAttribute="leading" id="4ca-Up-cUF"/>
-                <constraint firstAttribute="bottom" secondItem="TSG-Uf-nBd" secondAttribute="bottom" id="G2D-NI-RT9"/>
-                <constraint firstItem="TSG-Uf-nBd" firstAttribute="top" secondItem="gTV-IL-0wX" secondAttribute="top" id="tp3-9J-RVU"/>
-            </constraints>
-            <connections>
-                <outlet property="storeImageView" destination="TSG-Uf-nBd" id="7bG-OO-h74"/>
-            </connections>
-            <point key="canvasLocation" x="88" y="-67"/>
-        </collectionViewCell>
-    </objects>
-</document>
